Renamed UnallocatedCaseVie to CaseView, as it has elements that are visible when either allocated or not. Moved waitForCaseLoad to CaseView from Dashboard page. Moved some mothods from BasePage to Dashboard, as they were dashboard specific, and it was making refactoring some higher level methods tricky/confusing. Refactored dashboard.getAndClaimCurrentCase. We had logic in it that was checking if we were already on the current case before trying to load it from dashboard. This was causing race condition failures in scenarios, as the check might happen whilst the browser was moving between case view and dashboard. This method should only be used where we know we will want to load the case from the dashboard and then claim the case, we shouldnt have complex checks that allow part of that to be skipped. We have seperate getCurrentCase and claimCurrentCase that we can use in places where only one of the actions is required, and that will make the script of each scenario clearer. This change might cause some scenarios to fail if we have incorrectly used getAndClaimCurrentCase when we only needed claimCurrentCase, but this is desierable, as we would want to identify where we have incorrectly used it.
