+++ src/test/java/com/hocs/test/glue/dcu/DispatchStepDefs.java
@@ -2,7 +2,7 @@ package com.hocs.test.glue.dcu;
 
 import com.hocs.test.pages.decs.BasePage;
 import com.hocs.test.pages.decs.Dashboard;
-import com.hocs.test.pages.decs.UnallocatedCaseView;
+import com.hocs.test.pages.decs.CaseView;
 import com.hocs.test.pages.dcu.Dispatch;
 import io.cucumber.java.en.When;
 import net.thucydides.core.annotations.Managed;
@@ -15,13 +15,13 @@ public class DispatchStepDefs extends BasePage {
 
     Dashboard dashboard;
 
-    UnallocatedCaseView unallocatedCaseView;
+    CaseView caseView;
 
     @When("I complete the dispatch stage")
     public void completeTheDispatchStage() {
         if (!dispatch.dispatchAcceptRadioButton.isVisible()) {
             dashboard.getCurrentCase();
-            safeClickOn(unallocatedCaseView.allocateToMeLink);
+            safeClickOn(caseView.allocateToMeLink);
         }
         safeClickOn(dispatch.dispatchAcceptRadioButton);
         safeClickOn(continueButton);
