+++ src/test/java/com/hocs/test/glue/dcu/MarkupStepDefs.java
@@ -7,7 +7,7 @@ import static net.serenitybdd.core.Serenity.setSessionVariable;
 
 import com.hocs.test.pages.decs.BasePage;
 import com.hocs.test.pages.decs.Dashboard;
-import com.hocs.test.pages.decs.UnallocatedCaseView;
+import com.hocs.test.pages.decs.CaseView;
 import com.hocs.test.pages.decs.Workstacks;
 import com.hocs.test.pages.dcu.InitialDraft;
 import com.hocs.test.pages.dcu.Markup;
@@ -31,13 +31,13 @@ public class MarkupStepDefs extends BasePage {
 
     QAResponse qaResponse;
 
-    UnallocatedCaseView unallocatedCaseView;
+    CaseView caseView;
 
     @When("I complete the Markup stage")
     public void completeTheMarkupStage() {
         if (!markup.policyResponseRadioButton.isVisible()) {
             dashboard.getCurrentCase();
-            safeClickOn(unallocatedCaseView.allocateToMeLink);
+            safeClickOn(caseView.allocateToMeLink);
         }
         markup.moveCaseFromMarkupToInitialDraft();
     }
@@ -124,7 +124,7 @@ public class MarkupStepDefs extends BasePage {
 
     @Then("the case should be found in the {string} team")
     public void theCaseShouldBeFoundInTheTeamTeam(String team) {
-        goToDashboard();
+        dashboard.goToDashboard();
         switch (team.toUpperCase()) {
             case "PUBLIC PROTECTION UNIT":
                 safeClickOn(dashboard.publicProtectionUnit);
