+++ src/test/java/com/hocs/test/glue/dcu/PrivateOfficeApprovalStepDefs.java
@@ -10,7 +10,7 @@ import com.hocs.test.pages.decs.BasePage;
 import com.hocs.test.pages.decs.Dashboard;
 import com.hocs.test.pages.decs.SummaryTab;
 import com.hocs.test.pages.decs.TimelineTab;
-import com.hocs.test.pages.decs.UnallocatedCaseView;
+import com.hocs.test.pages.decs.CaseView;
 import com.hocs.test.pages.dcu.AccordionDCU;
 import com.hocs.test.pages.dcu.PrivateOfficeApproval;
 import io.cucumber.java.en.And;
@@ -26,7 +26,7 @@ public class PrivateOfficeApprovalStepDefs extends BasePage {
 
     AccordionDCU accordionDCU;
 
-    UnallocatedCaseView unallocatedCaseView;
+    CaseView caseView;
 
     SummaryTab summaryTab;
 
@@ -40,7 +40,7 @@ public class PrivateOfficeApprovalStepDefs extends BasePage {
             case "DTEN":
                 if (!privateOfficeApproval.privateOfficeAcceptRadioButton.isVisible()) {
                     dashboard.getCurrentCase();
-                    safeClickOn(unallocatedCaseView.allocateToMeLink);
+                    safeClickOn(caseView.allocateToMeLink);
                 }
                 safeClickOn(privateOfficeApproval.privateOfficeAcceptRadioButton);
                 safeClickOn(continueButton);
