+++ src/test/java/com/hocs/test/glue/dcu/QAResponseStepDefs.java
@@ -2,7 +2,7 @@ package com.hocs.test.glue.dcu;
 
 import com.hocs.test.pages.decs.BasePage;
 import com.hocs.test.pages.decs.Dashboard;
-import com.hocs.test.pages.decs.UnallocatedCaseView;
+import com.hocs.test.pages.decs.CaseView;
 import com.hocs.test.pages.decs.Workstacks;
 import com.hocs.test.pages.dcu.QAResponse;
 import io.cucumber.java.en.And;
@@ -17,13 +17,13 @@ public class QAResponseStepDefs extends BasePage {
 
     Workstacks workstacks;
 
-    UnallocatedCaseView unallocatedCaseView;
+    CaseView caseView;
 
     @When("I complete the QA response stage")
     public void completeQAResponseStage() {
         if (!qaResponse.QAAcceptRadioButton.isVisible()) {
             dashboard.getCurrentCase();
-            safeClickOn(unallocatedCaseView.allocateToMeLink);
+            safeClickOn(caseView.allocateToMeLink);
         }
         safeClickOn(qaResponse.QAAcceptRadioButton);
         System.out.println("Finished QA Response, returning to home page.");
