+++ src/test/java/com/hocs/test/glue/decs/CaseDataStepDefs.java
@@ -4,15 +4,13 @@ import com.hocs.test.pages.decs.BasePage;
 import com.hocs.test.pages.decs.Dashboard;
 import com.hocs.test.pages.decs.RecordCaseData;
 import com.hocs.test.pages.decs.SummaryTab;
-import com.hocs.test.pages.decs.UnallocatedCaseView;
+import com.hocs.test.pages.decs.CaseView;
 import com.hocs.test.pages.decs.Workstacks;
 import io.cucumber.java.en.And;
 
 public class CaseDataStepDefs extends BasePage {
 
     RecordCaseData recordCaseData;
-
-    UnallocatedCaseView unallocatedCaseView;
 
     SummaryTab summaryTab;
 
@@ -26,7 +24,7 @@ public class CaseDataStepDefs extends BasePage {
             summaryTab.selectSummaryTab();
             summaryTab.assertSummaryContainsExpectedValueForGivenHeader("User", getCurrentUser().getUsername());
             String assignedTeam = summaryTab.getSummaryTabValueForGivenHeader("Team");
-            goToDashboard();
+            dashboard.goToDashboard();
             dashboard.selectWorkstackByTeamName(assignedTeam);
             workstacks.unallocateSelectedCase(getCurrentCaseReference());
             workstacks.selectSpecificCaseReferenceLink(getCurrentCaseReference());
