+++ src/test/java/com/hocs/test/glue/decs/CreateCaseStepDefs.java
@@ -12,7 +12,7 @@ import com.hocs.test.pages.decs.CreateCase_SuccessPage;
 import com.hocs.test.pages.decs.Dashboard;
 import com.hocs.test.pages.decs.Documents;
 import com.hocs.test.pages.decs.SummaryTab;
-import com.hocs.test.pages.decs.UnallocatedCaseView;
+import com.hocs.test.pages.decs.CaseView;
 import com.hocs.test.pages.decs.Workstacks;
 import com.hocs.test.pages.comp.Registration;
 import com.hocs.test.pages.dcu.DataInput;
@@ -53,7 +53,7 @@ public class CreateCaseStepDefs extends BasePage {
 
     Workstacks workstacks;
 
-    UnallocatedCaseView unallocatedCaseView;
+    CaseView caseView;
 
     Creation creation;
 
@@ -80,7 +80,7 @@ public class CreateCaseStepDefs extends BasePage {
     @Given("I create a single {string} case and return to the dashboard")
     public void createACaseTypeSpecificCase(String caseType) {
         createCase.createCSCaseOfType(caseType.toUpperCase());
-        goToDashboard();
+        dashboard.goToDashboard();
     }
 
     @When("I create a {string} case {string} a document")
@@ -252,7 +252,7 @@ public class CreateCaseStepDefs extends BasePage {
     public void iCreateACaseWithAsTheCorrespondent(String caseType, String correspondent) {
         createCase.createCSCaseOfType(caseType.toUpperCase());
         createCaseSuccessPage.goToCaseFromSuccessfulCreationScreen();
-        safeClickOn(unallocatedCaseView.allocateToMeLink);
+        safeClickOn(caseView.allocateToMeLink);
         dataInput.completeDataInputStageWithMPCorrespondent(correspondent);
     }
 
@@ -286,33 +286,33 @@ public class CreateCaseStepDefs extends BasePage {
             default:
                 pendingStep(caseType + " is not defined within " + getMethodName());
         }
-        unallocatedCaseView.allocateToUserByVisibleText(user.getAllocationText());
+        caseView.allocateToUserByVisibleText(user.getAllocationText());
         setSessionVariable("selectedUser").to(user);
     }
 
     @And("I create and claim a MPAM case with {string} as the Urgency level and {string} as the Reference Type")
     public void iCreateAndClaimAMPAMCaseWithAsTheUrgencyLevelAndAsTheReferenceType(String urgency, String refType) {
         iCreateAMPAMCaseWithAsTheUrgencyLevelAndAsTheReferenceType(urgency, refType);
-        waitForDashboard();
+        dashboard.waitForDashboard();
         dashboard.getAndClaimCurrentCase();
     }
 
     @And("I create a MPAM case with {string} as the Urgency level and {string} as the Reference Type")
     public void iCreateAMPAMCaseWithAsTheUrgencyLevelAndAsTheReferenceType(String urgency, String refType) {
         createCase.createCSCaseOfType("MPAM");
-        goToDashboard();
+        dashboard.goToDashboard();
         dashboard.getAndClaimCurrentCase();
         creation.moveCaseWithSpecifiedUrgencyAndRefTypeToTriageStage(urgency, refType);
-        waitForDashboard();
+        dashboard.waitForDashboard();
     }
 
     @And("I create a Ministerial MPAM case with {string} as the Ministerial Sign Off Team and move it to the Triage stage")
     public void iCreateAMinisterialMPAMCaseWithAsTheMinisterialSignOffTeam(String signOffTeam) {
         createCase.createCSCaseOfType("MPAM");
-        goToDashboard();
+        dashboard.goToDashboard();
         dashboard.getAndClaimCurrentCase();
         creation.moveCaseWithSpecificMinisterialSignOffTeamToTriageStage(signOffTeam);
-        waitForDashboard();
+        dashboard.waitForDashboard();
     }
 
     @And("I create a {string} case with {string} as its {string}")
@@ -331,7 +331,7 @@ public class CreateCaseStepDefs extends BasePage {
                         break;
                     case "MEMBER OF PARLIAMENT NAME":
                         createCase.createCSCaseOfType("MIN");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         dataInput.completeDataInputStageWithMPCorrespondent(infoValue);
                         break;
@@ -339,10 +339,10 @@ public class CreateCaseStepDefs extends BasePage {
                     case "CORRESPONDENT POSTCODE":
                     case "CORRESPONDENT EMAIL ADDRESS":
                         createCase.createCSCaseOfType("MIN");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         dataInput.completeDataInputStageWithPublicCorrespondent();
-                        waitForDashboard();
+                        dashboard.waitForDashboard();
                         break;
                     case "TOPIC":
                         iCreateACaseWithAsIts("DCU", "Gordon Freeman", "Public Correspondent Name");
@@ -357,7 +357,7 @@ public class CreateCaseStepDefs extends BasePage {
                         break;
                     case "HOME SECRETARY INTEREST":
                         createCase.createCSCaseOfType("MIN");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         if (infoValue.equalsIgnoreCase("YES")) {
                             dataInput.completeDataInputStageSpecifyingHomeSecInterest(true);
@@ -374,29 +374,29 @@ public class CreateCaseStepDefs extends BasePage {
                     case "CASE REFERENCE":
                     case "ACTIVE CASES ONLY":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         break;
                     case "REFERENCE TYPE":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         creation.moveCaseWithSpecifiedBusinessAreaAndRefTypeToTriageStage("UKVI", infoValue);
                         break;
                     case "MINISTERIAL SIGN OFF TEAM":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         creation.moveCaseWithSpecificMinisterialSignOffTeamToTriageStage(infoValue);
                         break;
                     case "MEMBER OF PARLIAMENT NAME":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         creation.moveCaseWithSpecifiedMPCorrespondentToTriageStage(infoValue);
                         break;
                     case "CORRESPONDENT REFERENCE NUMBER":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         creation.moveCaseWithCorrespondentReferenceNumber(infoValue);
                         break;
@@ -408,7 +408,7 @@ public class CreateCaseStepDefs extends BasePage {
                         break;
                     case "CAMPAIGN":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         creation.moveCaseFromCreationToTriage();
                         dashboard.getAndClaimCurrentCase();
@@ -416,14 +416,14 @@ public class CreateCaseStepDefs extends BasePage {
                         break;
                     case "PUBLIC CORRESPONDENT NAME":
                         createCase.createCSCaseOfType("MPAM");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         creation.triggerMPCorrespondentIsMandatoryScreen();
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         break;
                     case "TELEPHONE SURGERY OFFICIAL ENGAGEMENT":
                         createCase.createCSCaseOfType("MTS");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         dashboard.getAndClaimCurrentCase();
                         mtsDataInput.completeDataInputStageAndCloseMTSCase();
                         break;
@@ -487,7 +487,7 @@ public class CreateCaseStepDefs extends BasePage {
                     case "TOPIC":
                     case "ACTIVE CASES ONLY":
                         foiCreateCase.createFOICase();
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         break;
                     case "RECEIVED ON OR AFTER":
                         dashboard.selectCreateSingleCaseLinkFromMenuBar();
@@ -505,7 +505,7 @@ public class CreateCaseStepDefs extends BasePage {
                         foiCreateCase.selectFOITopic("Animal alternatives (3Rs)");
                         foiCreateCase.enterRequestQuestion();
                         clickTheButton("Submit");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         break;
                     case "RECEIVED ON OR BEFORE":
                         dashboard.selectCreateSingleCaseLinkFromMenuBar();
@@ -523,7 +523,7 @@ public class CreateCaseStepDefs extends BasePage {
                         foiCreateCase.selectFOITopic("Animal alternatives (3Rs)");
                         foiCreateCase.enterRequestQuestion();
                         clickTheButton("Submit");
-                        goToDashboard();
+                        dashboard.goToDashboard();
                         break;
                     default:
                         pendingStep(infoType + " is not defined within " + getMethodName());
