+++ src/test/java/com/hocs/test/glue/decs/NavigationStepDefs.java
@@ -54,8 +54,8 @@ public class NavigationStepDefs extends BasePage {
 
     @And ("I click to view( the case/claim in) the {string} workstack")
     public void iClickToViewTheWorkstack(String workstackIdentifier) {
-        if (!onDashboard()) {
-            goToDashboard();
+        if (!dashboard.onDashboard()) {
+            dashboard.goToDashboard();
         }
         if (workstackIdentifier.equalsIgnoreCase("My Cases")) {
             dashboard.selectMyCases();
