+++ src/test/java/com/hocs/test/glue/ukvi/DraftStepDefs.java
@@ -4,6 +4,7 @@ import static jnr.posix.util.MethodName.getMethodName;
 import static net.serenitybdd.core.Serenity.pendingStep;
 
 import com.hocs.test.pages.decs.BasePage;
+import com.hocs.test.pages.decs.Dashboard;
 import com.hocs.test.pages.ukvi.Draft;
 import com.hocs.test.pages.ukvi.MPAMMultipleContributions;
 import io.cucumber.java.en.And;
@@ -14,6 +15,8 @@ public class DraftStepDefs extends BasePage {
     Draft draft;
 
     MPAMMultipleContributions MPAMMultipleContributions;
+
+    Dashboard dashboard;
 
     @And("I move a Official case from Draft to Dispatch bypassing QA")
     public void moveBRefCaseFromDraftToDispatch() {
@@ -42,7 +45,7 @@ public class DraftStepDefs extends BasePage {
             default:
                 pendingStep(action + " is not defined within " + getMethodName());
         }
-        waitForDashboard();
+        dashboard.waitForDashboard();
     }
 
     @When("I take the Draft \\(On Hold) case off hold")
